+++ client/client_test.go
@@ -15,8 +15,6 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/service/organizations"
-	orgTypes "github.com/aws/aws-sdk-go-v2/service/organizations/types"
 	"github.com/aws/aws-sdk-go-v2/service/sts"
 	stsTypes "github.com/aws/aws-sdk-go-v2/service/sts/types"
 )
@@ -300,93 +298,6 @@ func Test_Configure(t *testing.T) {
 		}
 		if a.AccessKeyID != tt.keyId {
 			t.Errorf("Case-%d failed: %+v", i, err)
-		}
-
-	}
-}
-
-func (m mockOrgClient) ListAccountsForParent(ctx context.Context, params *organizations.ListAccountsForParentInput, optFns ...func(*organizations.Options)) (*organizations.ListAccountsForParentOutput, error) {
-	return m.listAccountsForParent(ctx, params, optFns...)
-}
-
-func (m mockOrgClient) ListAccounts(ctx context.Context, params *organizations.ListAccountsInput, optFns ...func(*organizations.Options)) (*organizations.ListAccountsOutput, error) {
-	return m.listAccounts(ctx, params, optFns...)
-}
-
-type mockOrgClient struct {
-	listAccountsForParent func(ctx context.Context, params *organizations.ListAccountsForParentInput, optFns ...func(*organizations.Options)) (*organizations.ListAccountsForParentOutput, error)
-	listAccounts          func(ctx context.Context, params *organizations.ListAccountsInput, optFns ...func(*organizations.Options)) (*organizations.ListAccountsOutput, error)
-}
-
-func Test_Org_Configure(t *testing.T) {
-	ctx := context.Background()
-	tests := []struct {
-		listAccountsForParent func(ctx context.Context, params *organizations.ListAccountsForParentInput, optFns ...func(*organizations.Options)) (*organizations.ListAccountsForParentOutput, error)
-		listAccounts          func(ctx context.Context, params *organizations.ListAccountsInput, optFns ...func(*organizations.Options)) (*organizations.ListAccountsOutput, error)
-		ous                   []string
-		accounts              []Account
-		err                   error
-		config                *Config
-	}{
-		{
-			listAccounts: func(ctx context.Context, params *organizations.ListAccountsInput, optFns ...func(*organizations.Options)) (*organizations.ListAccountsOutput, error) {
-				return &organizations.ListAccountsOutput{
-					Accounts: []orgTypes.Account{},
-				}, nil
-			},
-			accounts: []Account{},
-			err:      nil,
-			config: &Config{
-				Organization: &AwsOrg{},
-			},
-		},
-		{
-			listAccountsForParent: func(ctx context.Context, params *organizations.ListAccountsForParentInput, optFns ...func(*organizations.Options)) (*organizations.ListAccountsForParentOutput, error) {
-				return &organizations.ListAccountsForParentOutput{
-					Accounts: []orgTypes.Account{
-						{
-							Id:     aws.String("012345678910"),
-							Status: orgTypes.AccountStatusActive,
-						},
-					},
-				}, nil
-			},
-			config: &Config{
-				Organization: &AwsOrg{
-					OrganizationUnits:    []string{"test-ou"},
-					ChildAccountRoleName: "test",
-				},
-			},
-			accounts: []Account{
-				{
-					ID:              "012345678910",
-					RoleARN:         "arn:aws:iam::012345678910:role/test",
-					RoleSessionName: "",
-					ExternalID:      "",
-					LocalProfile:    "",
-					Regions:         []string{},
-					source:          "org",
-				},
-			},
-			err: nil,
-		},
-	}
-
-	for _, test := range tests {
-		api := mockOrgClient{
-			listAccountsForParent: test.listAccountsForParent,
-			listAccounts:          test.listAccounts,
-		}
-		resp, err := loadAccounts(ctx, test.config, api)
-		respDiff := cmp.Diff(resp, test.accounts, cmpopts.IgnoreUnexported(Account{}), cmpopts.EquateEmpty())
-
-		if respDiff != "" {
-			t.Fatal(respDiff)
-		}
-		errDiff := cmp.Diff(err, test.err)
-
-		if errDiff != "" {
-			t.Fatal(errDiff)
 		}
 
 	}
