+++ client/organizations.go
@@ -1,0 +1,115 @@
+package client
+
+import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/organizations"
+	orgTypes "github.com/aws/aws-sdk-go-v2/service/organizations/types"
+	"github.com/aws/aws-sdk-go-v2/service/sts"
+	"github.com/hashicorp/go-hclog"
+)
+
+// Parses org configuration and grabs the appropriate accounts
+func loadOrgAccounts(ctx context.Context, logger hclog.Logger, awsConfig *Config) ([]Account, AssumeRoleAPIClient, error) {
+
+	// If user doesn't specify any configs for admin account instantiate default values
+	adminAccount := Account{
+		AccountName:  "Default-Admin-Account",
+		LocalProfile: "",
+	}
+	if awsConfig.Organization.AdminAccount != nil {
+		adminAccount = *awsConfig.Organization.AdminAccount
+	}
+	awsCfg, err := configureAwsClient(ctx, logger, awsConfig, adminAccount, nil)
+	if err != nil {
+		return nil, nil, err
+	}
+	svc := organizations.NewFromConfig(awsCfg)
+	accounts, err := loadAccounts(ctx, awsConfig, svc)
+	if err != nil {
+		return nil, nil, err
+	}
+	return accounts, sts.NewFromConfig(awsCfg), err
+
+}
+
+// Load accounts from the appropriate endpoint as well as normalizing response
+func loadAccounts(ctx context.Context, awsConfig *Config, accountsApi OrganizationsClient) ([]Account, error) {
+
+	var rawAccounts []orgTypes.Account
+	var err error
+	if len(awsConfig.Organization.OrganizationUnits) > 0 {
+		rawAccounts, err = getOUAccounts(ctx, accountsApi, awsConfig.Organization.OrganizationUnits)
+	} else {
+		rawAccounts, err = getAllAccounts(ctx, accountsApi)
+	}
+
+	if err != nil {
+		return []Account{}, err
+	}
+	accounts := make([]Account, 0)
+	for _, account := range rawAccounts {
+		// Only load Active accounts
+		if account.Status != orgTypes.AccountStatusActive {
+			continue
+		}
+		accounts = append(accounts, Account{
+			ID:              *account.Id,
+			RoleARN:         GenerateResourceARN("iam", "role", awsConfig.Organization.ChildAccountRoleName, "", *account.Id),
+			RoleSessionName: awsConfig.Organization.ChildAccountRoleSessionName,
+			ExternalID:      awsConfig.Organization.ChildAccountExternalID,
+			LocalProfile:    awsConfig.Organization.AdminAccount.LocalProfile,
+			Regions:         awsConfig.Organization.ChildAccountRegions,
+			source:          "org",
+		})
+	}
+	return accounts, err
+}
+
+// Get Accounts for specific Organizational Units
+func getOUAccounts(ctx context.Context, accountsApi OrganizationsClient, ous []string) ([]orgTypes.Account, error) {
+	var rawAccounts []orgTypes.Account
+
+	for _, ou := range ous {
+		var paginationToken *string
+		for {
+			resp, err := accountsApi.ListAccountsForParent(ctx, &organizations.ListAccountsForParentInput{
+				NextToken: paginationToken,
+				ParentId:  aws.String(ou),
+			})
+			if err != nil {
+				return nil, err
+			}
+			rawAccounts = append(rawAccounts, resp.Accounts...)
+			if resp.NextToken == nil {
+				break
+			}
+			paginationToken = resp.NextToken
+		}
+	}
+
+	return rawAccounts, nil
+}
+
+// Get All accounts in a specific organization
+func getAllAccounts(ctx context.Context, accountsApi OrganizationsClient) ([]orgTypes.Account, error) {
+	var rawAccounts []orgTypes.Account
+	var paginationToken *string
+
+	for {
+		resp, err := accountsApi.ListAccounts(ctx, &organizations.ListAccountsInput{
+			NextToken: paginationToken,
+		})
+		if err != nil {
+			return nil, err
+		}
+		rawAccounts = append(rawAccounts, resp.Accounts...)
+		if resp.NextToken == nil {
+			break
+		}
+		paginationToken = resp.NextToken
+	}
+	return rawAccounts, nil
+
+}
