+++ resources/services/organizations/accounts.go
@@ -84,7 +84,7 @@ func fetchOrganizationsAccounts(ctx context.Context, meta schema.ClientMeta, par
 			switch ae.ErrorCode() {
 			case "AWSOrganizationsNotInUseException", "AccessDeniedException":
 				// This is going to happen most probably due to account not being the root organizational account
-				// so it's better to ignore it completly as it happens basically on every account
+				// so it's better to ignore it completely as it happens basically on every account
 				// otherwise it screws up with dev experience and with our tests
 				meta.Logger().Warn("account is probably not the root organization account https://docs.aws.amazon.com/organizations/latest/APIReference/API_ListAccounts.html")
 				return nil
